HydroSigKMeans - hydro sig creation and clustering
Self-contained script to process a series of hydrodynamic signatures. The user needs only to download the dependencies for plotting the rose diagram and sorting file names, which are freely available on the internet (links included in the script). Define the file path where a series of raw measurement .mat files are located, and run the script. Afterwards, the hydrodynamic signatures are created and saved as PNG images. Furthermore, the script applies k-means clusterting and plots the elbow analysis results to determine the optimal number of k clusters.